+++ miden/src/cli/prove.rs
@@ -64,12 +64,20 @@ pub struct ProveCmd {
     /// Enable tracing to monitor execution of the VM
     #[clap(short = 't', long = "trace")]
     trace: bool,
+
+    /// Disable debug instructions (release mode)
+    #[clap(long = "release")]
+    release: bool,
 }
 
 impl ProveCmd {
     pub fn get_proof_options(&self) -> Result<ProvingOptions, ExecutionOptionsError> {
-        let exec_options =
-            ExecutionOptions::new(Some(self.max_cycles), self.expected_cycles, self.trace, false)?;
+        let exec_options = ExecutionOptions::new(
+            Some(self.max_cycles),
+            self.expected_cycles,
+            self.trace,
+            !self.release,
+        )?;
         Ok(match self.security.as_str() {
             "96bits" => {
                 if self.rpx {
@@ -111,7 +119,7 @@ impl ProveCmd {
                 get_masp_program(&self.program_file)?,
                 Arc::new(DefaultSourceManager::default()) as Arc<dyn SourceManager>,
             ),
-            "masm" => get_masm_program(&self.program_file, &libraries)?,
+            "masm" => get_masm_program(&self.program_file, &libraries, !self.release)?,
             _ => return Err(Report::msg("The provided file must have a .masm or .masp extension")),
         };