+++ miden/src/tools/mod.rs
@@ -54,7 +54,7 @@ impl Analyze {
                 get_masp_program(&self.program_file)?,
                 Arc::new(DefaultSourceManager::default()) as Arc<dyn SourceManager>,
             ),
-            "masm" => get_masm_program(&self.program_file, &libraries)?,
+            "masm" => get_masm_program(&self.program_file, &libraries, true)?,
             _ => return Err(Report::msg("The provided file must have a .masm or .masp extension")),
         };